scripts/pipeline/requirements.txt
# ── Core HTTP / async ────────────────────────────────────────────────────────
requests>=2.31.0
aiohttp>=3.9.0
aiofiles>=23.2.0

# ── Environment / config ──────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Image processing ──────────────────────────────────────────────────────────
Pillow>=10.0.0
numpy>=1.24.0

# ── CLI progress ──────────────────────────────────────────────────────────────
tqdm>=4.66.0
rich>=13.7.0

# ── Background removal (Stage 3) ──────────────────────────────────────────────
# CPU runtime — swap for onnxruntime-gpu if you have a CUDA GPU
rembg>=2.0.50
onnxruntime>=1.16.0
# onnxruntime-gpu>=1.16.0  # uncomment and remove onnxruntime line for GPU

# ── QC (Stage 4) ─────────────────────────────────────────────────────────────
imagehash>=4.3.1
Flask>=3.0.0

# sqlite3 is stdlib — no pip install needed

# ── Optional: alternative 3D / image APIs ────────────────────────────────────
# replicate>=0.22.0   # Replicate client
# stability-sdk>=0.8.4  # Stability AI SDK (stage2_generator uses raw HTTP)
